requirements.txt
streamlit
google-generativeai
langchain
pandas
plotly
python-dotenv